+++ src/main.spec.ts
@@ -14,7 +14,6 @@ import {
 } from "../test/faker";
 import { stub } from "../test/mocks";
 import { mockedCypressEventEmitter } from "../test/util";
-import globalContext, { PluginContext } from "./context";
 import { configureXrayPlugin, syncFeatureFile } from "./main";
 import type {
     CypressFailedRunResult,
@@ -22,6 +21,7 @@ import type {
     PluginConfigOptions,
 } from "./models/cypress";
 import type { PluginOptions } from "./models/plugin";
+import globalContext, { PluginContext } from "./plugin/context";
 import cypressXrayPlugin from "./plugin/cypress-xray-plugin";
 import { dedent } from "./util/dedent";
 import { LOG } from "./util/logging";
