+++ src/main.ts
@@ -1,14 +1,14 @@
 import path from "path";
-import globalContext from "./context";
-import type { PluginTaskParameterType } from "./cypress/tasks";
-import { CypressTaskListener } from "./cypress/tasks";
 import type { CypressFailedRunResult, CypressRunResult } from "./models/cypress";
 import type {
     CypressXrayPluginOptions,
     InternalCypressXrayPluginOptions,
     InternalPluginOptions,
 } from "./models/plugin";
+import globalContext from "./plugin/context";
 import cypressXrayPlugin from "./plugin/cypress-xray-plugin";
+import type { PluginTaskParameterType } from "./tasks/tasks";
+import { CypressTaskListener } from "./tasks/tasks";
 import { dedent } from "./util/dedent";
 import { getOrCall } from "./util/functions";
 import { HELP } from "./util/help";
