+++ src/plugin/context.ts
@@ -1,17 +1,16 @@
 import axios from "axios";
-import type { HttpCredentials } from "./client/authentication/credentials";
+import type { HttpCredentials } from "../client/authentication/credentials";
 import {
     BasicAuthCredentials,
     JwtCredentials,
     PatCredentials,
-} from "./client/authentication/credentials";
-import { AxiosRestClient } from "./client/https/requests";
-import { JiraClientCloud } from "./client/jira/jira-client-cloud";
-import { JiraClientServer } from "./client/jira/jira-client-server";
-import { XrayClientCloud } from "./client/xray/xray-client-cloud";
-import { XrayClientServer } from "./client/xray/xray-client-server";
-import { ENV_NAMES } from "./env";
-import type { ObjectLike, ScreenshotDetails } from "./models/cypress";
+} from "../client/authentication/credentials";
+import { AxiosRestClient } from "../client/https/requests";
+import { JiraClientCloud } from "../client/jira/jira-client-cloud";
+import { JiraClientServer } from "../client/jira/jira-client-server";
+import { XrayClientCloud } from "../client/xray/xray-client-cloud";
+import { XrayClientServer } from "../client/xray/xray-client-server";
+import type { ObjectLike, ScreenshotDetails } from "../models/cypress";
 import type {
     ClientCombination,
     CypressXrayPluginOptions,
@@ -23,15 +22,16 @@ import type {
     InternalPluginOptions,
     InternalXrayOptions,
     PluginEvent,
-} from "./models/plugin";
-import type { XrayEvidenceItem } from "./models/xray/import-test-execution-results";
-import { dedent } from "./util/dedent";
-import type { CucumberPreprocessorArgs, CucumberPreprocessorExports } from "./util/dependencies";
-import dependencies from "./util/dependencies";
-import { errorMessage } from "./util/errors";
-import { HELP } from "./util/help";
-import { CapturingLogger, LOG } from "./util/logging";
-import { asArrayOfStrings, asBoolean, asObject, asString, parse } from "./util/parsing";
+} from "../models/plugin";
+import type { XrayEvidenceItem } from "../models/xray/import-test-execution-results";
+import { dedent } from "../util/dedent";
+import type { CucumberPreprocessorArgs, CucumberPreprocessorExports } from "../util/dependencies";
+import dependencies from "../util/dependencies";
+import { ENV_NAMES } from "../util/env";
+import { errorMessage } from "../util/errors";
+import { HELP } from "../util/help";
+import { CapturingLogger, LOG } from "../util/logging";
+import { asArrayOfStrings, asBoolean, asObject, asString, parse } from "../util/parsing";
 
 export class PluginContext {
     private readonly internalOptions: InternalCypressXrayPluginOptions;
