+++ src/plugin/plugin-phases.ts
@@ -16,11 +16,6 @@ import type {
     HasAddEvidenceEndpoint,
     HasGetTestRunEndpoint,
 } from "../client/xray/xray-client-server";
-import type {
-    EvidenceCollection,
-    IterationParameterCollection,
-    PluginEventEmitter,
-} from "../context";
 import type { PluginConfigOptions, ScreenshotDetails } from "../models/cypress";
 import type {
     InternalCucumberOptions,
@@ -32,6 +27,11 @@ import type {
 import type { MultipartInfo } from "../models/xray/requests/import-execution-multipart-info";
 import { dedent } from "../util/dedent";
 import type { Logger } from "../util/logging";
+import type {
+    EvidenceCollection,
+    IterationParameterCollection,
+    PluginEventEmitter,
+} from "./context";
 import type { MinimalCypressRunResult } from "./cypress-xray-plugin";
 import featureFileProcessing from "./feature-file-processing/feature-file-processing";
 import featureFileUpload from "./feature-file-upload/feature-file-upload";
