+++ src/tasks/tasks.spec.ts
@@ -3,8 +3,8 @@ import path from "node:path";
 import process from "node:process";
 import { describe, it } from "node:test";
 import { getMockedCypress } from "../../test/mocks";
-import { SimpleEvidenceCollection, SimpleIterationParameterCollection } from "../context";
 import type { RequestOptions } from "../models/cypress";
+import { SimpleEvidenceCollection, SimpleIterationParameterCollection } from "../plugin/context";
 import { dedent } from "../util/dedent";
 import { LOG } from "../util/logging";
 import * as tasks from "./tasks";
